requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
Flask==3.1.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
greenlet==3.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
mysql-connector-python==9.3.0
numpy==2.2.5
packaging==25.0
propcache==0.3.1
psycopg==3.0.1
PuLP==3.1.1
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.3
setuptools==58.2.0
soupsieve==2.7
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
